Allow console to use cert from system certbot
In case we alredy have certbot for the system' servives, we do not need
to run yet another certbot to get the console's cert: we can simply
reuse what's already available. This patch move the certbot
configuration for the cloud from cloud-post to cloud (thus making sure
we have the certbot before we start the console), and change the certbot
configuation of the console to not run certbot again if we already have
the cert.